add routing with some temp hardcoding
